Refactor test fixtures into conftest.py and update tests to use fixtures
- Moved create_user and create_clothing_item fixtures into conftest.py for better reusability.
- Updated test cases in test_user_crud.py and test_clothing_item_crud.py to use the shared fixtures.
- Ensured that the test cases correctly assert using '_id' instead of 'id'.

Closes #5
